requirements.txt
numpy
faiss-cpu
sentence-transformers
python-docx
PyPDF2
requests